API/requirements.txt
aniso8601==8.0.0
click==8.1.3
Flask==2.3.3
# Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
itsdangerous==2.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==3.1
llama-cpp-python==0.3.8
scikit-fuzzy
numpy
scipy
packaging
networkx
# ctransformers
torch
requests 
python-dotenv 
openai